requirements_flask.txt
# libs for flask app
chroma-hnswlib==0.7.6
chromadb==0.5.11
Flask==2.2.2
Flask-Cors==3.0.10
# Flask-JWT-Extended==4.4.4
groq==0.11.0
langchain==0.3.2
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.9
langchain-groq==0.2.0
langchain-huggingface==0.1.0
py-healthcheck==1.10.1
# PyJWT==2.4.0
sentence-transformers==3.1.1
Werkzeug==2.2.2